Rename custom widgets module. Only check filters if spatial filters are present to avoid confusion.
